add MatchTOFParams

+++ Detectors/TOF/workflowIO/src/DigitReaderSpec.cxx
@@ -35,7 +35,7 @@ void DigitReader::init(InitContext& ic)
   LOG(debug) << "Init Digit reader!";
   auto filename = o2::utils::Str::concat_string(o2::utils::Str::rectifyDirectory(ic.options().get<std::string>("input-dir")),
                                                 ic.options().get<std::string>("tof-digit-infile"));
-  mDelayInMuSec1TF = atoi(ic.options().get<std::string>("delay-1st-tf").c_str()) * 1E6;
+  mDelayInMuSec1TF = atof(ic.options().get<std::string>("delay-1st-tf").c_str()) * 1E6;
 
   mFile.reset(TFile::Open(filename.c_str()));
   if (!mFile->IsOpen()) {
